Second step of structural refactoring: Final proposition.
