+++ src/client/app/fassung/fassung-blaettern/fassung-blaettern.component.ts
@@ -2,13 +2,13 @@
  * Created by Reto Baumgartner (rfbaumgartner) on 05.07.17.
  */
 
-import {Component, Input} from '@angular/core';
+import { Component, Input } from '@angular/core';
 
 @Component({
   moduleId: module.id,
   selector: 'rae-fassung-blaettern',
   templateUrl: 'fassung-blaettern.component.html',
-  styleUrls: ['fassung-blaettern.component.css']
+  styleUrls: [ 'fassung-blaettern.component.css' ]
 })
 export class FassungBlaetternComponent {
   @Input() idOfPrev: string;
