+++ src/client/app/fassung/fassung-steckbrief/fassung-steckbrief.component.ts
