+++ src/client/app/konvolut/konvolut-grid/konvolut-grid.component.css
