+++ src/client/app/konvolut/konvolut-steckbrief/konvolut-steckbrief.component.css
