+++ src/client/app/konvolut/super-konvolut.component.ts
@@ -9,7 +9,7 @@ import { ActivatedRoute, Params, Router } from '@angular/router';
 import 'rxjs/add/operator/catch';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/switchMap';
-import { ExtendedSearch, KnoraProperty } from '../../shared/utilities/knora-api-params';
+import { ExtendedSearch, KnoraProperty } from '../shared/utilities/knora-api-params';
 
 @Component({
   moduleId: module.id,
