--- src/client/app/konvolute/shared/konstituierter-text.component.ts
@@ -1,12 +1,0 @@
-/**
- * Created by Sebastian Schüpbach (sebastian.schuepbach@unibas.ch) on 6/7/17.
- */
-
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'rae-konstituierter-text',
-  template: '<h1>Konstituierter Text</h1>'
-})
-export class KonstituierterTextComponent {
-}
