+++ src/client/app/register/register-navigation/register-navigation.component.ts
@@ -1,7 +1,7 @@
 /**
  * Created by Reto Baumgartner (rfbaumgartner) on 07.07.17.
  */
-import {Component, Input} from '@angular/core';
+import { Component, Input } from '@angular/core';
 
 import 'rxjs/add/operator/catch';
 import 'rxjs/add/operator/map';
@@ -9,10 +9,11 @@ import 'rxjs/add/operator/switchMap';
 
 @Component({
   moduleId: module.id,
-  selector: 'rae-register-beschreibung',
-  templateUrl: 'register-beschreibung.component.html'
+  selector: 'rae-register-navigation',
+  templateUrl: 'register-navigation.component.html',
+  styleUrls: [ 'register-navigation.component.css' ]
 })
-export class RegisterBeschreibungComponent {
+export class RegisterNavigationComponent {
 
   @Input() selectedTab: string;
