+++ src/client/app/register/register.module.ts
@@ -4,16 +4,13 @@
 
 import { NgModule } from '@angular/core';
 import { HttpModule } from '@angular/http';
-import {
-  MdButtonModule, MdListModule,
-} from '@angular/material';
+import { MdButtonModule, MdListModule } from '@angular/material';
 import { BrowserModule } from '@angular/platform-browser';
 import { RegisterComponent } from './register.component';
 import { RegisterRoutingModule } from './register-routing.module';
 import { RegisterNavigationComponent } from './register-navigation/register-navigation.component';
 import { RegisterBeschreibungComponent } from './register-beschreibung/register-beschreibung.component';
 import { RegisterTitelregisterComponent } from './titelregister/register-titelregister.component';
-
 
 
 @NgModule({
