+++ src/client/app/register/titelregister/register-titelregister.component.ts
@@ -8,7 +8,7 @@ import { ActivatedRoute, Params, Router } from '@angular/router';
 
 import 'rxjs/add/operator/catch';
 import 'rxjs/add/operator/map';
-import { FulltextSearch } from '../../../shared/utilities/knora-api-params';
+import { FulltextSearch } from '../../shared/utilities/knora-api-params';
 
 @Component({
   moduleId: module.id,
