+++ src/client/app/shared/registerspalte/registerspalte.module.ts
