+++ src/client/app/synopse/synopse.component.ts
